% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CAN_govcsv}
\alias{CAN_govcsv}
\title{Canada Public Health Coronavirus Cumulative Case Totals}
\format{
A data frame where each row represents a reporting date
}
\source{
\url{https://www.canada.ca/en/public-health/services/diseases/2019-novel-coronavirus-infection.html}
}
\usage{
CAN_govcsv
}
\description{
A dataset containing the data reported online by the Canadian government,
which is published as a .csv file, titled covid19.csv. The data is
transformed into wide format.
}
\keyword{datasets}
