% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ECDC_owid}
\alias{ECDC_owid}
\title{European Centre for Disease Control cases and deaths data}
\format{
A data frame with a row for each date
}
\source{
\url{https://ourworldindata.org/coronavirus-source-data}
}
\usage{
ECDC_owid
}
\description{
A dataset containing the data reported by the ECDC, as curated and released
by Our World in Data. This object is in long format as presented by OWiD.
}
\keyword{datasets}
