% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_city}
\alias{map_city}
\title{a sf object of all the Chinese prefecture cities}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 371 rows and 4 columns.}
\usage{
map_city
}
\description{
a sf object of all the Chinese prefecture cities
}
\keyword{datasets}
