% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_provice}
\alias{map_provice}
\title{a sf object of all the Chinese provinces}
\format{An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 4 columns.}
\usage{
map_provice
}
\description{
a sf object of all the Chinese provinces
}
\keyword{datasets}
