% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ON_linelist}
\alias{ON_linelist}
\title{Ontario Ministry of Health Status of COVID-19 cases in Ontario}
\format{
A data frame where each row confirmed case
}
\source{
\url{https://data.ontario.ca/dataset/confirmed-positive-cases-of-covid-19-in-ontario}
}
\usage{
ON_linelist
}
\description{
A dataset containing the data reported online by the Ontario Ministry of 
Health in its .csv file conposcovidloc.csv. Variable definitions from the
Ministry's provided data dictionary (conposcovidloc_data_dictionary.xlsx).
}
\keyword{datasets}
