% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ON_mohreports}
\alias{ON_mohreports}
\title{Ontario Ministry of Health Epidemiologic Summaries (PDF reports)}
\format{A data frame where each row represents a report}
\source{
\url{https://www.ontario.ca/page/2019-novel-coronavirus}
}
\usage{
ON_mohreports
}
\description{
A dataset containing the data reported online by the Ontario Ministry of 
Health in its Epidemiologic Summary reports.
}
\keyword{datasets}
