% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ON_status}
\alias{ON_status}
\title{Ontario Ministry of Health Status of COVID-19 cases in Ontario}
\format{
A data frame where each row represents a date's data
 \describe{
    \item{date}{Reporting date of Ontario COVID-19 daily summary}
    \item{negative}{Patients tested negative based on testing performed at both Public Health Ontario (PHO) Laboratory and National Microbiology Lab (NML)}
    \item{positive}{Number of active confirmed positive COVID-19 cases in Ontario as measured on "Reported Date"}
    \item{presumptive}{Patients tested positive at Public Health Ontario (PHO)  Laboratory, but not yet tested by National Microbiology Lab (NML)}
    \item{resolved}{Cumulative number of resolved cases as measured on "Reported Date"}
    \item{deaths}{Cumulative number of fatalities of people with COVID-19 as measured on "Reported Date"}
    \item{cases}{Cumulative number of confirmed cases reported to date in the Integrated Public Health Information System (iPHIS) by Ontario’s 34 Public Health Units (PHUs), including resolved and deceased cases as measured on "Reported Date"}
    \item{tested_patients}{Cumulative number of tests submitted to the COVID-19 Clinical Lab network (including PHO, hospital, academic and community laboratories) as measured on "Reported Date"}
    \item{tests_last_day}{Number of tests processed and completed during the day before the "Reported Date" (the 24-hour period from 12:00 a.m. – 11:59 p.m.)}
    \item{current_investigation}{Number of tests pending results as measured on "Reported Date"}
    \item{current_hospitalized}{Number of patients currently hospitalized with COVID-19 as measured on "Reported Date"}
    \item{current_ICU}{Number of patients currently in an Intensive Care Unit (ICU) with COVID-19 as measured on "Reported Date"}
    \item{current_ventilator}{Number of patients currently in an Intensive Care Unit (ICU), on a ventilator with COVID-19 as measured on "Reported Date"}
 }
}
\source{
\url{https://data.ontario.ca/dataset/status-of-covid-19-cases-in-ontario}
}
\usage{
ON_status
}
\description{
A dataset containing the data reported online by the Ontario Ministry of 
Health in its .csv file covidtesting.csv. Variable definitions from the
Ministry's provided data dictionary (covidtesting_data_dictionary.xlsx).
}
\keyword{datasets}
