% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sarscov2_ecdc_2019}
\alias{sarscov2_ecdc_2019}
\title{European Centre for Disease Control cases and deaths data}
\format{
A data frame with a row for each date
}
\source{
\url{https://ourworldindata.org/coronavirus-source-data}
}
\usage{
sarscov2_ecdc_2019
}
\description{
A dataset containing the data reported by the ECDC, as curated and released
by Our World in Data. This object is in wide format with ISO 3-letter country
code variable names.
}
\keyword{datasets}
