% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WHO_SR}
\alias{WHO_SR}
\title{World Health Organization situation reports data}
\format{A data frame with a row for each situation report}
\source{
\url{https://www.who.int/emergencies/diseases/novel-coronavirus-2019/situation-reports}
}
\usage{
WHO_SR
}
\description{
A dataset containing the data reported in the WHO 2019 novel
coronavirus situation reports.
}
\keyword{datasets}
