% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ON_cumulative}
\alias{ON_cumulative}
\title{Ontario Ministry of Health Coronavirus Data}
\format{A data frame where each row represents a new Situation Report
 \describe{
    \item{URL}{Internet Archive snapshot of URL for source data}
    \item{LastUpdated}{The date the of the data reported on the webpage}
    \item{Negative}{Negative tests.}
    \item{UnderInvestigation}{Number of current pending tests.}
    \item{PresumptivePositive}{Awaiting confirmation at national laboratory.}
    \item{ConfirmedPositive}{"Patient still testing positive and has not had two consecutive negative results greater than 24 hours apart"}
    \item{Resolved}{"Patient is no longer infectious based on two consecutive negative tests performed"}
    \item{TotalTested}{"Total number of patients approved for COVID-19 testing to date"}
 }}
\source{
\url{https://www.ontario.ca/page/2019-novel-coronavirus}
}
\usage{
ON_cumulative
}
\description{
A dataset containing the data reported online by the Ontario Ministry of 
Health. For verification purposes, the data is retrieved from the Internet
Archive web archives, with the URL of the snapshot provided.
}
\keyword{datasets}
