% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ON_cumulative}
\alias{ON_cumulative}
\title{Ontario Ministry of Health Coronavirus Cumulative Case Totals}
\format{A data frame where each row represents a reporting time
 \describe{
    \item{URL}{Internet Archive snapshot of URL for source data}
    \item{AsOf}{The results are updated to this date}
    \item{LastUpdated}{The date the data were reported on the webpage}
    \item{Negative}{Negative tests.}
    \item{UnderInvestigation}{Number of current pending tests.}
    \item{PresumptivePositive}{Awaiting confirmation at national laboratory.}
    \item{Deceased}{Confirmed deceased cases.}
    \item{ConfirmedPositive}{"Patient still testing positive and has not had two consecutive negative results greater than 24 hours apart"}
    \item{Cases}{"Number of confirmed cases reported to date in iPHIS by 34 Public Health Units, including resolved and deceased cases"}
    \item{Resolved}{"Patient is no longer infectious based on two consecutive negative tests performed"}
    \item{TotalTested}{"Total number of patients approved for COVID-19 testing to date"}
    \item{Male}{"Male cases"}
    \item{Female}{"Female cases"}
    \item{X19andunder}{"Cases age 19 and under"} 
    \item{X20to64}{"Cases age 20 to 64"}
    \item{X65andover}{"Cases age 65 and over"}
    \item{X20to39}{"Cases age 20 to 39"}
    \item{X40to59}{"Cases age 40 to 59"}
    \item{X60to79}{"Cases age 60 to 79"}
    \item{X80andover}{"Cases age 80 and over"}
    \item{Hospitalized}{"Cases currently hospitalized"}
    \item{ICU}{"Cases currently in ICU"}
    \item{Ventilator}{"Cases currently in ICU and on ventilator"}
 }}
\source{
\url{https://www.ontario.ca/page/2019-novel-coronavirus}
}
\usage{
ON_cumulative
}
\description{
A dataset containing the data reported online by the Ontario Ministry of 
Health. For verification purposes, the data is retrieved from the Internet
Archive web archives, with the URL of the snapshot provided.
}
\keyword{datasets}
