% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{sf_prefecture_city}
\alias{sf_prefecture_city}
\title{sf_prefecture_city}
\usage{
sf_prefecture_city(month, day, hour)
}
\arguments{
\item{month}{a integer}

\item{day}{a integer}

\item{hour}{a integer}
}
\value{
a sf object
}
\description{
get the 2019-nCov data at the prefecture city level
}
\examples{
\dontrun{
# get the 2019-nCov data at the prefecture city level at the time: 20:00, January 29.
sf_prefecture_city(month = 1, day = 29, hour = 20)
}
}
