% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{sf_provice}
\alias{sf_provice}
\title{sf_provice}
\usage{
sf_provice(month, day, hour)
}
\arguments{
\item{month}{a integer}

\item{day}{a integer}

\item{hour}{a integer}
}
\value{
a sf object
}
\description{
get the 2019-nCov data at provincial level
}
\examples{
\dontrun{
get the 2019-nCov data at provincial level at the time: 20:00, January 29.
sf_provice(month = 1, day = 29, hour = 20)
}
}
